+++ src/state.rs
@@ -27,18 +27,18 @@ use std::time::UNIX_EPOCH;
 use syntect::highlighting::{Theme, ThemeSet};
 
 #[cfg(target_family = "unix")]
-use std::os::unix::fs::PermissionsExt;
-#[cfg(target_family = "unix")]
-use std::os::unix::fs::MetadataExt;
-#[cfg(target_family = "unix")]
 use nix::sys::stat::Mode;
 #[cfg(target_family = "unix")]
 use nix::unistd::{Gid, Uid};
+#[cfg(target_family = "unix")]
+use std::os::unix::fs::MetadataExt;
+#[cfg(target_family = "unix")]
+use std::os::unix::fs::PermissionsExt;
 
-pub const FELIX: &str = "felix";
 pub const BEGINNING_ROW: u16 = 3;
 pub const EMPTY_WARNING: &str = "Are you sure to empty the trash directory? (if yes: y)";
 const TIME_PREFIX: usize = 11;
+const BASE32: &[u8; 32] = b"0123456789ABCDEFGHJKMNPQRSTVWXYZ";
 
 #[derive(Debug)]
 pub struct State {
@@ -53,6 +53,7 @@ pub struct State {
     pub p_memo: Vec<StateMemo>,
     pub keyword: Option<String>,
     pub layout: Layout,
+    pub is_ro: bool,
 }
 
 #[derive(Default, Debug, PartialEq, Eq, PartialOrd, Ord, Clone)]
@@ -88,31 +89,19 @@ impl Default for FileType {
 
 impl State {
     /// Initialize the state of the app.
-    pub fn new(p: &std::path::Path, session_path: &std::path::Path) -> Result<Self, FxError> {
-        let config = match read_config(p) {
+    pub fn new(session_path: &std::path::Path) -> Result<Self, FxError> {
+        //Read config file.
+        //Use default configuration if the file does not exist or cannot be read.
+        let config = read_config_or_default();
+        let config = match config {
             Ok(c) => c,
             Err(e) => {
-                eprintln!("Cannot read the config file properly.\nError: {}\nDo you want to use the default config? [press Enter to continue]", e);
-                enter_raw_mode();
-                loop {
-                    match crossterm::event::read()? {
-                        Event::Key(KeyEvent { code, .. }) => match code {
-                            KeyCode::Enter => break,
-                            _ => {
-                                leave_raw_mode();
-                                return Err(FxError::Yaml("Exit the app.".to_owned()));
-                            }
-                        },
-                        _ => {
-                            continue;
-                        }
-                    }
-                }
-                leave_raw_mode();
+                eprintln!("Cannot read the config file properly.\nError: {}\nfelix launches with default configuration.", e);
                 Config::default()
             }
         };
-        let session = read_session(session_path)?;
+
+        let session = read_session(session_path);
         let (original_column, original_row) = terminal_size()?;
 
         // Return error if terminal size may cause panic
@@ -179,6 +168,7 @@ impl State {
             c_memo: Vec::new(),
             p_memo: Vec::new(),
             keyword: None,
+            is_ro: false,
         })
     }
 
@@ -864,15 +854,12 @@ impl State {
             header_space -= current_dir.len();
         }
 
-        #[cfg(target_family = "unix")]
         // If without the write permission, print [RO].
-        if let Ok(false) = has_write_permission(&self.current_dir) {
-            if header_space > 5 {
-                set_color(&TermColor::ForeGround(&Colorname::Red));
-                print!(" [RO]");
-                reset_color();
-                header_space -= 5;
-            }
+        if self.is_ro && header_space > 5 {
+            set_color(&TermColor::ForeGround(&Colorname::Red));
+            print!(" [RO]");
+            reset_color();
+            header_space -= 5;
         }
 
         //If .git directory exists, get the branch information and print it.
@@ -1085,6 +1072,33 @@ impl State {
         }
     }
 
+    /// Creates temp file for directory. Works like touch, but with randomized suffix
+    #[allow(dead_code)]
+    pub fn create_temp(&mut self, is_dir: bool) -> Result<PathBuf, FxError> {
+        let mut new_name = self.current_dir.join(".tmp");
+        if new_name.exists() {
+            let mut nanos = std::time::SystemTime::now()
+                .duration_since(UNIX_EPOCH)
+                .unwrap()
+                .subsec_nanos();
+            let encoded: &mut [u8] = &mut [0, 0, 0, 0, 0];
+            for i in 0..5 {
+                let v = (nanos & 0x1f) as usize;
+                encoded[4 - i] = BASE32[v];
+                nanos >>= 5;
+            }
+            new_name = self
+                .current_dir
+                .join(format!(".tmp_{}", String::from_utf8(encoded.to_vec())?))
+        }
+        if is_dir {
+            std::fs::create_dir(new_name.clone())?;
+        } else {
+            std::fs::File::create(new_name.clone())?;
+        }
+        Ok(new_name)
+    }
+
     //Show help
     pub fn show_help(&self, mut screen: &Stdout) -> Result<(), FxError> {
         clear_all();
@@ -1164,6 +1178,10 @@ impl State {
 
     pub fn chdir(&mut self, p: &std::path::Path, mv: Move) -> Result<(), FxError> {
         std::env::set_current_dir(p)?;
+        self.is_ro = match has_write_permission(p) {
+            Ok(b) => !b,
+            Err(_) => false,
+        };
         match mv {
             Move::Up => {
                 // Push current state to c_memo
@@ -1665,7 +1683,7 @@ fn choose_theme(dt: &DefaultTheme) -> Theme {
 // Currently available in unix only.
 // TODO: Use this function to determine if deleting items can be done in the first place?
 #[cfg(target_family = "unix")]
-fn has_write_permission(path: &std::path::Path) -> Result<bool, FxError> {
+pub fn has_write_permission(path: &std::path::Path) -> Result<bool, FxError> {
     let metadata = std::fs::metadata(path)?;
     let mode = metadata.mode();
     if mode == 0 {
@@ -1689,6 +1707,12 @@ fn has_write_permission(path: &std::path::Path) -> Result<bool, FxError> {
     }
 }
 
+// Currently on non-unix OS, this always returns true.
+#[cfg(not(target_family = "unix"))]
+pub fn has_write_permission(_path: &std::path::Path) -> Result<bool, FxError> {
+    Ok(true)
+}
+
 #[cfg(all(target_family = "unix", not(target_os = "macos")))]
 fn in_groups(gid: u32) -> bool {
     if let Ok(groups) = nix::unistd::getgroups() {
@@ -1706,16 +1730,16 @@ fn in_groups(gid: u32) -> bool {
 }
 
 #[cfg(target_os = "macos")]
-fn in_groups(gid: u32) -> bool {
+fn in_groups(_gid: u32) -> bool {
     false
 }
 
 #[cfg(test)]
 mod tests {
+    use super::*;
+
     use devtimer::run_benchmark;
     use rayon::prelude::{IntoParallelIterator, ParallelIterator};
-
-    use super::*;
 
     fn bench_update1() -> Result<(), FxError> {
         let mut dir_v = Vec::new();
@@ -1754,8 +1778,19 @@ mod tests {
 
     #[test]
     fn test_has_write_permission() {
+        // chmod to 444 and check if it's read-only
         let p = std::path::PathBuf::from("./testfiles/permission_test");
+        let _status = std::process::Command::new("chmod")
+            .args(["444", "./testfiles/permission_test"])
+            .status()
+            .unwrap();
         assert!(!has_write_permission(p.as_path()).unwrap());
+        let _status = std::process::Command::new("chmod")
+            .args(["755", "./testfiles/permission_test"])
+            .status()
+            .unwrap();
+
+        // Test the home directory, which should pass
         let home_dir = dirs::home_dir().unwrap();
         assert!(has_write_permission(&home_dir).unwrap());
     }